add support for any model
