+++ .gitignore
@@ -1,3 +1,6 @@
+# project specific
+example/*
+
 # See https://help.github.com/ignore-files/ for more about ignoring files.
 
 # Byte-compiled / optimized / DLL files
