+++ gpt_engineer/models.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 from enum import Enum
-
+from typing import Any, Dict, List, Tuple, Optional
 
 class Step(Enum):
     CLARIFY = 'clarify'
